Use f64 end-to-end to match Sinumerik REAL precision
Sinumerik program data of type REAL is a 64-bit IEEE double (NC
programming manual 4.1.3.6); R parameters and system variables are
TYPE_DOUBLE, and comparison commands are specified with a relative
equality of 1e-12 - none of which is representable in f32 (~7
significant digits, relative epsilon ~1.2e-7).

Migrate Value, State (axes/symbol_table/translation) and the whole
expression evaluator from f32 to f64. DIV now divides the REAL
operands and truncates the result (manual 4.1.3.1: 7 DIV 4.1 = 1)
instead of truncating the operands through integer casts, which also
removes the bogus division-by-zero error for divisors below 1.
Output DataFrame columns become Float64 natively instead of being
cast from f32-quantized values, so emitted coordinates are now
bit-exact for decimal literals.

Closes #22

+++ src/output.rs
@@ -14,7 +14,7 @@ use std::collections::{HashMap, HashSet};
 /// A single typed column. Values are optional: `None` is a null cell.
 #[derive(Debug, Clone, PartialEq)]
 pub enum Column {
-    Float(Vec<Option<f32>>),
+    Float(Vec<Option<f64>>),
     Int(Vec<Option<i64>>),
     Str(Vec<Option<String>>),
     StrList(Vec<Option<Vec<String>>>),
@@ -198,7 +198,7 @@ fn build_column(name: &str, rows: &[&HashMap<String, Value>]) -> Column {
         .iter()
         .map(|r| match r.get(name) {
             Some(Value::Float(f)) => Some(*f),
-            Some(Value::Str(s)) => s.parse::<f32>().ok(),
+            Some(Value::Str(s)) => s.parse::<f64>().ok(),
             _ => None,
         })
         .collect();
